Add files via upload
All files must be within one collected file. When downloading, all JavaScript (js) files must be saved in a file called 'js'. When downloading 'style.css' it must be saved in a file called 'css'. Finally when downloading the 'readme.txt' file and the 'fancy-cushion.png' then they must be saved in a folder called 'img.'